refactor(ws): moved socketMessage to own file
